pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["github_token_updater"]

[project]
name = "github_token_updater"
version = "0.0.1"
description = "Update a secret across multiple repositories"
authors = [{ name = "Joshua Sundance Bailey" }]
readme = "README.md"
dependencies = [
    "PyNaCl",
    "requests",
]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["github", "secrets"]
[project.urls]
Homepage = "https://github.com/joshuasundance-swca/github_token_updater"


[project.scripts]
github_token_updater = "github_token_updater.__main__:main"

[project.optional-dependencies]
dev = [
    "bumpver",
    "pre-commit",
    "pytest",
]

[tool.bumpver]
current_version = "0.0.1"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
commit = true
tag_message = "{new_version}"
tag_scope = "default"
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"github_token_updater/__init__.py" = ['__version__ = "{version}"']
